+++ google2/CHANGELOG.md
@@ -8,6 +8,7 @@ Changed:
 - `pollOperation` in `GoogleComputeService` now returns `Stream[F, Operation]`
 - bug fix in `deleteBucket`
 - Don't throw Not Found when listing objects for empty bucket in `deleteBucket`
+- Fix a bug in `GoogleDataprocService` where region is not set properly
 
 Added:
 - `GKEInterpreter`, `GKEService`, `KubernetesService`, and `KubernetesInterpreter`
@@ -17,8 +18,9 @@ Added:
 - add `deleteBucket` to `GoogleStorageService`
 - add optional `credentials` parameter to `GoogleStorageService.getBlob`
 - `{create,get,delete}Nodepool` to `GKEService`
+- Add `getClusterInstances` and `getClusterError`
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.8-e08439a"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.8-TRAVIS-REPLACE-ME"`
 
 ## 0.7
 Changed: